feat: Fix hero text section image to text ratio
This took some playing around but finally figured out
that min-width: 0% would allow the image to properly
flex to be equal with the hero text message. They
now take up equal space even when the height of the section
changes.

This is important so that I can play with the appearance of
the layout without messing up the ratio of the image
to the hero text. The issue is that they aren't aligned
top and bottom like I would have hoped, but this will have to
suffice.